added grandularity on iOS
