+++ packages/health/example/.flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"device_id","path":"/Users/tnni/.pub-cache/hosted/pub.dartlang.org/device_id-0.2.0/","dependencies":[]},{"name":"health","path":"/Users/tnni/GitHub/flutter-plugins/packages/health/","dependencies":["device_id"]}],"android":[{"name":"device_id","path":"/Users/tnni/.pub-cache/hosted/pub.dartlang.org/device_id-0.2.0/","dependencies":[]},{"name":"health","path":"/Users/tnni/GitHub/flutter-plugins/packages/health/","dependencies":["device_id"]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"device_id","dependencies":[]},{"name":"health","dependencies":["device_id"]}],"date_created":"2020-07-28 17:19:38.757580","version":"1.17.5"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"device_id","path":"/Users/tnni/.pub-cache/hosted/pub.dartlang.org/device_id-0.2.0/","dependencies":[]},{"name":"health","path":"/Users/tnni/GitHub/flutter-plugins/packages/health/","dependencies":["device_id"]}],"android":[{"name":"device_id","path":"/Users/tnni/.pub-cache/hosted/pub.dartlang.org/device_id-0.2.0/","dependencies":[]},{"name":"health","path":"/Users/tnni/GitHub/flutter-plugins/packages/health/","dependencies":["device_id"]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"device_id","dependencies":[]},{"name":"health","dependencies":["device_id"]}],"date_created":"2020-08-06 13:11:42.139185","version":"1.17.5"}
